Address PR comments, and rename module for consistency
