+++ src/chat-model-handler.ts
